MaterialSystem: fix incorrect member variable type
Should be unsigned int 16, not 32. Member gets accessed with uint16 instructions. Structure is now alligned correctly with the engine, and its size also matches now.

+++ src/materialsystem/cmaterialglue.h
@@ -57,7 +57,7 @@ struct MaterialGlue_s
 	uint32 materialFlags2;
 
 	MaterialRenderParams_s renderParams[MATERIAL_RENDER_PARAMS_COUNT];
-	uint32 numAnimationFrames;
+	uint16 numAnimationFrames;
 	MaterialShaderType_e materialType;
 	uint8 uberBufferFlags;
